optimization def _create() of factories
