+++ fake_data_factories/company_user_factories.py
@@ -7,11 +7,10 @@ import factory
 from termcolor import cprint
 
 from constants import FAKER_USER_COUNT
-from fake_data_factories.base_user_factory import BaseUserFactory, password_helper
+from fake_data_factories.base_user_factory import BaseUserFactory
 from fake_data_factories.company_factories import CompanyFactory
 from src.database.alembic_models import UserTabit
 from src.database.sc_db_session import sc_session
-from src.logger import fake_db_logger
 
 
 class PositionEnum(str, Enum):
@@ -45,20 +44,6 @@ class CompanyUserFactory(BaseUserFactory):
     async def company_id(self):
         company = await CompanyFactory.create()
         return company.id
-
-    @classmethod
-    def _create(cls, model_class, *args, **kwargs):
-        password = kwargs.pop('password', None)
-        if password:
-            kwargs['hashed_password'] = password_helper.hash(password)
-        else:
-            password = factory.Faker('password').evaluate(None, None, {'locale': 'ru_RU'})
-            kwargs['hashed_password'] = password_helper.hash(password)
-        fake_db_logger.info(
-            f'Сотрудник компании c id={kwargs.get("company_id")}: '
-            f'{kwargs.get("email")}, пасс: {password}'
-        )
-        return super()._create(model_class, *args, **kwargs)
 
 
 async def create_company_users(count: int = FAKER_USER_COUNT, **kwargs) -> None:
